reducers split
