+++ src/components/burger-constructor/reducers/constructor-reducer.js
@@ -13,18 +13,6 @@ const initialState = {
 
 export function constructorReducer(state = initialState, action) {
   switch (action.type) {
-    case actions.CREATE_ORDER_REQUEST: {
-      return {...state, createdOrderRequest: true};
-    }
-
-    case actions.CREATE_ORDER_SUCCESS: {
-      return {...state, createdOrderRequest: false, orderNumber: action.orderNumber};
-    }
-
-    case actions.CREATE_ORDER_FAIL: {
-      return {...state, createdOrderRequest: false, createOrderFail: true};
-    }
-
     case actions.ADD_BUN: {
       return {...state, bun: {...action.bun}};
     }
