+++ src/components/ingredient-details/actions/ingredient-details-actions.js
@@ -1,0 +1,2 @@
+export const INGREDIENT_MODAL_OPEN = 'INGREDIENT_MODAL_OPEN';
+export const INGREDIENT_MODAL_CLOSED = 'INGREDIENT_MODAL_CLOSED';
